src/backend_api/requirements.txt
flask==3.1.1
kafka-python==2.2.10
pymongo==4.13.0
